tests(transfer): move forwarding tests to separate file. (#6568)
